refactor(app): add models and services folders for cards
