+++ src/app/models/card.models/abstract-card.model.ts
