+++ src/app/models/card.models/artifact-card.model.ts
@@ -1,5 +1,5 @@
 import { ObtainableCard } from "./abstract-card.model";
-import { ARTIFACT_TAX } from "../constants/prices";
+import { ARTIFACT_TAX } from "../../constants/prices";
 
 export class ArtifactCard extends ObtainableCard {
   readonly tax = ARTIFACT_TAX;
