+++ src/app/models/card.models/base-card.model.ts
