+++ src/app/models/card.models/fate-card.model.ts
