+++ src/app/models/card.models/trap-card.model.ts
