+++ src/app/services/card.services/fate-card.service.ts
@@ -1,6 +1,6 @@
-import { FateCard } from "../models/fate-card.model";
-import { Player } from "../models/player.model";
-import { BAD_OR_GOOD_SALARY, BONUS_SALARY } from "../constants/prices";
+import { FateCard } from "../../models/card.models/fate-card.model";
+import { Player } from "../../models/player.model";
+import { BAD_OR_GOOD_SALARY, BONUS_SALARY } from "../../constants/prices";
 
 export const badChampionEarn = (player: Player) => {
   if(player.goodChampion) player.giveMoney(BAD_OR_GOOD_SALARY);
