+++ src/app/services/card.services/trap-card.service.ts
@@ -1,6 +1,6 @@
-import { Player } from "../models/player.model";
-import { TRAP_TAX } from "../constants/prices";
-import { PLAYER_ATTACKED_STEPS, PLAYER_IS_CHASED_STEPS } from "../constants/game-constants";
+import { Player } from "../../models/player.model";
+import { TRAP_TAX } from "../../constants/prices";
+import { PLAYER_ATTACKED_STEPS, PLAYER_IS_CHASED_STEPS } from "../../constants/game-constants";
 
 export const loseMoneyAndMoveBackward = (player: Player) => {
   player.giveMoney(TRAP_TAX);
